requirements.txt
# LLM and AI Framework
crewai>=0.28.0
crewai-tools>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Validation and Schemas
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0

# Optional: For local LLM support
# llama-cpp-python>=0.2.0
# transformers>=4.30.0

